+++ src/main/java/com/safefood/dto/CartDto.java
@@ -9,12 +9,12 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class CartDto {
     private String id;
-    private int code;
+    private Long code;
     private int quantity;
     private String name;
     private String imgPath;
 
-    public CartDto(String id, int code, int quantity) {
+    public CartDto(String id, Long code, int quantity) {
         this.id = id;
         this.code = code;
         this.quantity = quantity;
