+++ src/main/java/com/safefood/dto/IntakeDto.java
@@ -10,12 +10,12 @@ import lombok.NoArgsConstructor;
 public class IntakeDto {
     private int intakeCode;
     private String id;
-    private int code;
+    private Long code;
     private int quantity;
     private String name;
     private String imgPath;
 
-    public IntakeDto(String id, int code, int quantity) {
+    public IntakeDto(String id, Long code, int quantity) {
         this.id = id;
         this.code = code;
         this.quantity = quantity;
